@@ -1,20 +1,22 @@
 module github.com/avelino/awesome-go
 
-go 1.17
+go 1.21
 
 require (
-	github.com/PuerkitoBio/goquery v1.8.0
+	github.com/PuerkitoBio/goquery v1.8.1
 	github.com/avelino/slugify v0.0.0-20180501145920-855f152bd774
-	github.com/gomarkdown/markdown v0.0.0-20211212230626-5af6ad2f47df
-	github.com/russross/blackfriday v1.6.0
-	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8
+	github.com/otiai10/copy v1.14.0
+	github.com/yuin/goldmark v1.6.0
+	golang.org/x/oauth2 v0.15.0
 )
 
 require (
 	github.com/andybalholm/cascadia v1.3.1 // indirect
-	github.com/golang/protobuf v1.4.2 // indirect
-	golang.org/x/net v0.0.0-20210916014120-12bc252f5db8 // indirect
-	golang.org/x/text v0.3.6 // indirect
-	google.golang.org/appengine v1.6.6 // indirect
-	google.golang.org/protobuf v1.25.0 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	golang.org/x/net v0.19.0 // indirect
+	golang.org/x/sync v0.3.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	google.golang.org/appengine v1.6.7 // indirect
+	google.golang.org/protobuf v1.33.0 // indirect
 )
